penalty added if taxi donot have enough fuel

+++ TaxiMeter-OO/TaxiMeter-OO/Trip.h
@@ -20,7 +20,7 @@
 -(float)calcTimeDifference;
 -(void)calcCost:(float) difference taxiObject:(Taxi *)taxi;
 -(float)absFloat:(float)f;
--(int)fact:(int)i;
+-(double)fact:(int)i;
 -(double)sine:(double)x;
 -(double) cosine:(double)x;
 -(double) tangent:(double)x;
